requirements.txt
tensorflow==2.5.0
plotly==5.1.0
numpy==1.19.5
streamlit==0.85.1
Pillow==8.3.1
flask==1.1.2
jinja2==2.11.3
werkzeug==1.0.1
